refactor(frontend): split admin fetch job into counter, expansion, preview, and meta services
