+++ frontend/src/main/java/world/willfrog/alphafrogmicro/frontend/kafka/FetchTaskStatusListener.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 import world.willfrog.alphafrogmicro.frontend.config.TaskProducerRabbitConfig;
 import world.willfrog.alphafrogmicro.common.dao.agent.AdminFetchTaskDao;
 import world.willfrog.alphafrogmicro.common.pojo.agent.AdminFetchTask;
-import world.willfrog.alphafrogmicro.frontend.service.AdminFetchJobService;
+import world.willfrog.alphafrogmicro.frontend.service.AdminFetchJobCounterService;
 import world.willfrog.alphafrogmicro.frontend.service.AdminFetchTaskService;
 import world.willfrog.alphafrogmicro.frontend.service.FetchTaskStatusService;
 
@@ -24,7 +24,7 @@ public class FetchTaskStatusListener {
 
     private final FetchTaskStatusService fetchTaskStatusService;
     private final AdminFetchTaskService adminFetchTaskService;
-    private final AdminFetchJobService adminFetchJobService;
+    private final AdminFetchJobCounterService adminFetchJobCounterService;
     private final AdminFetchTaskDao adminFetchTaskDao;
 
     @RabbitListener(queues = TaskProducerRabbitConfig.FETCH_RESULT_QUEUE)
@@ -95,7 +95,7 @@ public class FetchTaskStatusListener {
             AdminFetchTask task = adminFetchTaskDao.getByTaskUuid(taskUuid);
             if (task != null && task.getJobUuid() != null) {
                 try {
-                    adminFetchJobService.refreshJobCounters(task.getJobUuid());
+                    adminFetchJobCounterService.refreshJobCounters(task.getJobUuid());
                 } catch (Exception e) {
                     log.error("Failed to refresh job counters for jobUuid={}", task.getJobUuid(), e);
                 }
